backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Data & ML
pandas==2.2.0
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.0
statsmodels==0.14.1

# NLP
nltk==3.8.1

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Report Generation
reportlab==4.1.0
openpyxl==3.1.2
jinja2==3.1.3

# Utilities
httpx==0.27.0
python-dotenv==1.0.1
uuid6==2024.1.12

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0
